LayoutTests/imported/w3c/web-platform-tests/css/css-conditional/container-queries/container-units-invalidation-expected.txt

Test
Test

FAIL cqi respond when selected container changes type (inline-size -> normal) assert_equals: expected "30px" but got "80px"
FAIL cqb respond when selected container changes type (size -> normal) assert_equals: expected "40px" but got "60px"
FAIL cqb respond when intermediate container changes type (inline-size -> size) assert_equals: expected "40px" but got "60px"
FAIL cqi respond when selected container changes inline-size assert_equals: expected "30px" but got "80px"
FAIL cqb respond when selected container changes block-size assert_equals: expected "40px" but got "60px"
